server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
mysql-connector-python==8.3.0
python-dotenv==1.0.0
PyJWT==2.8.0